DjangoAdmin/pyproject.toml
[project]
name = "djangoadmin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cloudinary>=1.44.1",
    "django>=5.2.7",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-unfold>=0.68.0",
    "djangorestframework>=3.16.1",
    "markdown>=3.9",
    "uvicorn>=0.37.0",
    "whitenoise>=6.11.0",
]
[tool.autopep8]
indent_size = 2